It started of as project to learn more about arrays, but it ended transforming into a project that manages a list of your "dream cars".
There is a lot more I would like to add to it and improve upon, This is basically my first draft.

The sql part was done with the help of chatGPT since I am not very familiar with it yet, I am excited to keep working on it and see how far i could take this.

It's not an innovative project by any means, but it's a fun way for me to learn more about programming and experiment a little.